es/requirements.txt
PyYAML~=5.1.1
elasticsearch~=8.13.0